requirements.txt
BeautifulSoup==3.2.1
Cython==0.20.1
Flask==0.10.1
Flask-WTF==0.9.5
Jinja2==2.7.2
MarkupSafe==0.19
MySQL-python==1.2.5
Pigments==1.6
Pillow==2.4.0
PyMySQL==0.6.2
PyYAML==3.11
Shapely==1.3.2
Unirest==1.1.6
WTForms==1.0.5
Werkzeug==0.9.4
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
fuzzywuzzy==0.2
geopy==0.99
gnureadline==6.3.3
gunicorn==18.0
ipython==2.1.0
itsdangerous==0.23
lxml==3.3.5
matplotlib==1.3.1
nose==1.3.3
numexpr==2.4
numpy==1.8.1
pandas==0.14.0
poster==0.8.1
protobuf==2.5.0
pygeocoder==1.2.4
pyparsing==2.0.2
python-Levenshtein==0.11.2
python-dateutil==2.2
python-google-places==0.10.1
pytz==2014.4
pyzmq==14.3.0
requests==2.3.0
scikit-image==0.10.0
scikit-learn==0.14.1
scipy==0.14.0
six==1.6.1
tables==3.1.1
tornado==3.2.2
wsgiref==0.1.2
